refactor: 크롤링 3회 반복 로직 삭제
