+++ hashtagmap-instagram-service/src/main/java/com/songpapeople/hashtagmap/service/CrawlerWithProxy.java
@@ -3,7 +3,6 @@ package com.songpapeople.hashtagmap.service;
 import com.songpapeople.hashtagmap.crawler.InstagramCrawler;
 import com.songpapeople.hashtagmap.exception.CrawlerException;
 import com.songpapeople.hashtagmap.exception.CrawlerExceptionStatus;
-import com.songpapeople.hashtagmap.exception.InstagramSchedulerException;
 import com.songpapeople.hashtagmap.place.domain.model.Place;
 import lombok.extern.slf4j.Slf4j;
 
@@ -11,8 +10,7 @@ import java.util.Optional;
 
 @Slf4j
 public class CrawlerWithProxy {
-    private static final int START_TRY_COUNT = 0;
-    private static final int MAX_TRY_COUNT = 3;
+    private static final String NOT_FOUND_EXCEPTION_CODE = CrawlerExceptionStatus.NOT_FOUND_URL.getStatusCode();
 
     private final ProxySetter proxySetter;
     private final InstagramCrawler instagramCrawler;
@@ -23,25 +21,15 @@ public class CrawlerWithProxy {
     }
 
     public Optional<CrawlingResult> crawlInstagram(Place place, String hashtagNameToCrawl) {
-        return crawlInstagram(place, hashtagNameToCrawl, START_TRY_COUNT);
-    }
-
-    private Optional<CrawlingResult> crawlInstagram(Place place, String hashtagNameToCrawl, int tryCount) {
-        if (tryCount >= MAX_TRY_COUNT) {
-            return Optional.empty();
-        }
         try {
             proxySetter.setProxy();
             return Optional.of(new CrawlingResult(instagramCrawler.crawler(hashtagNameToCrawl), place));
         } catch (CrawlerException e) {
             log.info("CrawlerException: {}", e.getMessage());
-            if (CrawlerExceptionStatus.NOT_FOUND_URL.getStatusCode().equals(e.getErrorCode())) {
+            if (NOT_FOUND_EXCEPTION_CODE.equals(e.getErrorCode())) {
                 return Optional.empty();
             }
-            return crawlInstagram(place, hashtagNameToCrawl, tryCount + 1);
-        } catch (InstagramSchedulerException e) {
-            log.info("InstagramSchedulerException: {}", e.getMessage());
-            return Optional.empty();
+            throw e;
         }
     }
 }
