+++ hashtagmap-instagram-service/src/main/java/com/songpapeople/hashtagmap/service/InstagramCrawlingService.java
@@ -4,13 +4,12 @@ import com.songpapeople.hashtagmap.blacklist.domain.model.BlackList;
 import com.songpapeople.hashtagmap.blacklist.domain.repsitory.BlackListRepository;
 import com.songpapeople.hashtagmap.crawler.InstagramCrawler;
 import com.songpapeople.hashtagmap.dto.CrawlingDto;
-import com.songpapeople.hashtagmap.exception.CrawlerException;
-import com.songpapeople.hashtagmap.exception.CrawlerExceptionStatus;
 import com.songpapeople.hashtagmap.instagram.domain.model.Instagram;
 import com.songpapeople.hashtagmap.instagram.domain.model.InstagramPost;
 import com.songpapeople.hashtagmap.instagram.domain.repository.InstagramRepository;
 import com.songpapeople.hashtagmap.instagram.domain.repository.instagramPost.InstagramPostRepository;
 import com.songpapeople.hashtagmap.place.domain.model.Place;
+import com.songpapeople.hashtagmap.proxy.ProxiesFactory;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -24,8 +23,6 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @Service
 public class InstagramCrawlingService {
-    private static final String NOT_FOUND_EXCEPTION_CODE = CrawlerExceptionStatus.NOT_FOUND_URL.getStatusCode();
-
     private final InstagramRepository instagramRepository;
     private final InstagramPostRepository instagramPostsRepository;
     private final BlackListRepository blackListRepository;
@@ -35,18 +32,9 @@ public class InstagramCrawlingService {
         if (isSkipPlace(place)) {
             return Optional.empty();
         }
-
         String hashtagNameToCrawl = findHashtagNameToCrawl(place);
-
-        try {
-            return Optional.of(new CrawlingResult(instagramCrawler.crawler(hashtagNameToCrawl), place));
-        } catch (CrawlerException e) {
-            log.info("CrawlerException: {}", e.getMessage());
-            if (NOT_FOUND_EXCEPTION_CODE.equals(e.getErrorCode())) {
-                return Optional.empty();
-            }
-            throw e;
-        }
+        CrawlerWithProxy crawlerWithProxy = new CrawlerWithProxy(new ProxySetter(ProxiesFactory.create()), instagramCrawler);
+        return crawlerWithProxy.crawlInstagram(place, hashtagNameToCrawl);
     }
 
     @Transactional
