+++ hashtagmap-instagram-service/src/test/java/com/songpapeople/hashtagmap/service/CrawlerWithProxyTest.java
@@ -1,10 +1,8 @@
 package com.songpapeople.hashtagmap.service;
 
 import com.songpapeople.hashtagmap.crawler.InstagramCrawler;
-import com.songpapeople.hashtagmap.dto.CrawlingDto;
 import com.songpapeople.hashtagmap.exception.CrawlerException;
 import com.songpapeople.hashtagmap.exception.CrawlerExceptionStatus;
-import com.songpapeople.hashtagmap.exception.InstagramSchedulerException;
 import com.songpapeople.hashtagmap.place.domain.model.Place;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -16,7 +14,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.*;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class CrawlerWithProxyTest {
@@ -33,42 +32,23 @@ class CrawlerWithProxyTest {
         crawlerWithProxy = new CrawlerWithProxy(proxySetter, instagramCrawler);
     }
 
-    @DisplayName("검색할 수 없는 Place 검색시 Optional.empty()반환")
+    @DisplayName("429 예외가 발생하면 예외를 반환한다.")
     @Test
-    void instagramCrawling() {
-        // given
+    void tooManyRequestExceptionTest() {
         Place place = Place.builder()
                 .placeName("잠실타로&사주")
                 .build();
-        when(instagramCrawler.crawler(place.getPlaceName())).thenThrow(InstagramSchedulerException.class);
+        CrawlerException exception = new CrawlerException(CrawlerExceptionStatus.TOO_MANY_REQUEST);
+        when(instagramCrawler.crawler(place.getPlaceName())).thenThrow(exception);
 
-        // when
-        Optional<CrawlingResult> actual = crawlerWithProxy.crawlInstagram(place, place.getPlaceName());
-
-        // then
-        assertThat(actual).isEqualTo(Optional.empty());
+        assertThatThrownBy(() -> crawlerWithProxy.crawlInstagram(place, place.getPlaceName()))
+                .isInstanceOf(exception.getClass())
+                .hasMessage(exception.getMessage());
     }
 
-    @DisplayName("크롤링 반복 횟수 확인 - CrawlerException 발생 (429,연결할 수 없을 때) 시 3회 실행")
+    @DisplayName("404 예외가 발생하면 빈 결과를 반환한다.")
     @Test
-    void crawlerExceptionTimesTest() {
-        // given
-        Place place = Place.builder()
-                .placeName("잠실타로&사주")
-                .build();
-        when(instagramCrawler.crawler(place.getPlaceName())).thenThrow(CrawlerException.class);
-
-        // when
-        crawlerWithProxy.crawlInstagram(place, place.getPlaceName());
-
-        // then
-        verify(instagramCrawler, times(3)).crawler(place.getPlaceName());
-    }
-
-    @DisplayName("크롤링 반복 횟수 확인 - 가게 검색 결과가 존재하지 않을 때 (404) 1회 실행")
-    @Test
-    void crawlerExceptionTimesTest_404() {
-        // given
+    void notFoundPlaceExceptionTest() {
         Place place = Place.builder()
                 .placeName("잠실타로&사주")
                 .build();
@@ -76,29 +56,9 @@ class CrawlerWithProxyTest {
         when(instagramCrawler.crawler(place.getPlaceName())).thenThrow(exception);
 
         // when
-        crawlerWithProxy.crawlInstagram(place, place.getPlaceName());
+        Optional<CrawlingResult> result = crawlerWithProxy.crawlInstagram(place, place.getPlaceName());
 
         // then
-        verify(instagramCrawler, times(1)).crawler(place.getPlaceName());
-    }
-
-    @DisplayName("크롤링 반복 횟수 확인 - InstagramSchedulerException 발생 시 1회 실행")
-    @Test
-    void instagramSchedulerExceptionTimesTest_404() {
-        // given
-        Place place = Place.builder()
-                .placeName("잠실타로&사주")
-                .build();
-
-        final int NOT_ENOUGH_HASHTAG_COUNT = 100;
-        CrawlingDto crawlingDto = CrawlingDto.of(place.getPlaceName(), String.valueOf(NOT_ENOUGH_HASHTAG_COUNT), null);
-
-        when(instagramCrawler.crawler(anyString())).thenReturn(crawlingDto);
-
-        // when
-        crawlerWithProxy.crawlInstagram(place, place.getPlaceName());
-
-        // then
-        verify(instagramCrawler, times(1)).crawler(place.getPlaceName());
+        assertThat(result.isPresent()).isFalse();
     }
 }
